requirements.txt
streamlit>=1.40.0
chromadb>=0.6.0
sentence-transformers>=3.0.0
google-generativeai>=0.8.0
pandas>=2.0.0
requests>=2.31.0
tenacity>=8.2.3
langdetect>=1.0.9
numpy>=1.26.0
pytest>=8.0.0
